requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
fastapi==0.115.12
h11==0.16.0
idna==3.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-multipart==0.0.20
python-jose==3.4.0
passlib==1.7.4
python-dotenv==1.1.0
psycopg2-binary==2.9.10
redis==6.0.0
setuptools==80.3.1
sniffio==1.3.1
starlette==0.46.2
sqlalchemy==2.0.40
types-cffi==1.17.0.20250326
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.3.0.20250505
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
